fix: compile error

+++ Assets/Scripts/Pet/ProjectilePetAttack.cs
@@ -4,19 +4,19 @@ using UnityEngine;
 
 public class ProjectilePetAttack : MonoBehaviour
 {
-    public float damage = 10f;
+    public int damage = 10;
     // Start is called before the first frame update
     void Start()
     {
-        
+
     }
 
     // Update is called once per frame
     void Update()
     {
-        
+
     }
-    
+
     void OnTriggerEnter(Collider other)
     {
         if (other.CompareTag("Enemy"))
@@ -28,8 +28,8 @@ public class ProjectilePetAttack : MonoBehaviour
             }
             Destroy(this.gameObject);
         }
-        
-        if(LayerMask.LayerToName(other.gameObject.layer) == "Ground")
+
+        if (LayerMask.LayerToName(other.gameObject.layer) == "Ground")
         {
             Destroy(this.gameObject);
         }
